Added the exact date. As it only had month and year
Could not find change that I could suggest. Learned a lot while looking at your code. There are still some gaps in the knowledge I gained but I am sure I will be able to understand it in due time.

The parallax effect in the home page is also very captivating. And the way you used AI to generate the image to use it was also a very innovative thought, I never though I could use AI in such a way.